+++ lib/log-types.d.ts
@@ -1,5 +1,9 @@
 export declare enum LogType {
     logcat = "logcat",
     bugreport = "bugreport",
-    server = "server"
+    server = "server",
+    driver = "driver",
+    syslog = "syslog",
+    crashlog = "crashlog",
+    performance = "performance"
 }